Add tests and refactor test structure
